requirements.txt
python-dateutil
aprs
requests